Unify CLI entrypoints under sdk
