+++ docs/agent_deployment.md
@@ -1,7 +1,7 @@
 # API based deployment
 
 Agents can be deployed via the MCP API. A simple client is implemented in
-`agentnn.deployment.agent_registry.AgentRegistry`. Use the CLI tool
-`cli/agentctl.py deploy AGENT.yaml` to upload a definition to a running
+`agentnn.deployment.agent_registry.AgentRegistry`. Use the unified CLI
+`python -m sdk.cli agent deploy AGENT.yaml` to upload a definition to a running
 MCP server.
