+++ sdk/cli/commands/session_runner.py
@@ -1,9 +1,4 @@
-"""Interactive CLI for running Agent-NN sessions."""
-
-from __future__ import annotations
-
 import json
-import time
 from pathlib import Path
 from typing import Optional
 
@@ -13,12 +8,13 @@ import yaml
 from agentnn.reasoning.context_reasoner import MajorityVoteReasoner
 from agentnn.session.session_manager import SessionManager
 from agentnn.storage import snapshot_store
+from ...client import AgentClient
 
-app = typer.Typer(help="Run and observe live agent sessions")
+session_app = typer.Typer(name="session", help="Session utilities")
 manager = SessionManager()
 
 
-@app.command()
+@session_app.command("start")
 def start(template: Optional[Path] = None) -> None:
     """Start a new session from optional YAML template."""
     sid = manager.create_session()
@@ -37,14 +33,14 @@ def start(template: Optional[Path] = None) -> None:
     typer.echo(json.dumps({"session_id": sid}))
 
 
-@app.command()
+@session_app.command("watch")
 def watch(session_id: str) -> None:
     """Print message history for a session."""
     session = manager.get_session(session_id)
     typer.echo(json.dumps(session.get("message_history", []), indent=2))
 
 
-@app.command()
+@session_app.command("vote")
 def vote(session_id: str, roles: str = "") -> None:
     """Aggregate last step results using majority vote."""
     session = manager.get_session(session_id)
@@ -60,14 +56,14 @@ def vote(session_id: str, roles: str = "") -> None:
     typer.echo(json.dumps({"decision": reasoner.decide()}))
 
 
-@app.command()
+@session_app.command("refine")
 def refine(session_id: str, message: str) -> None:
     """Run another task with all agents."""
     manager.run_task(session_id, message)
     typer.echo("ok")
 
 
-@app.command()
+@session_app.command("snapshot")
 def snapshot(session_id: str) -> None:
     """Save snapshot of session state."""
     session = manager.get_session(session_id) or {}
@@ -75,12 +71,19 @@ def snapshot(session_id: str) -> None:
     typer.echo(json.dumps({"snapshot_id": snap}))
 
 
-@app.command()
+@session_app.command("restore")
 def restore(snapshot_id: str) -> None:
     """Restore a snapshot and return new session id."""
     sid = snapshot_store.restore_snapshot(snapshot_id)
     typer.echo(json.dumps({"session_id": sid}))
 
 
-if __name__ == "__main__":
-    app()
+@session_app.command("budget")
+def session_budget(session_id: str) -> None:
+    """Show consumed tokens for a session."""
+    client = AgentClient()
+    data = client.get_session_history(session_id)
+    tokens = 0
+    for ctx in data.get("context", []):
+        tokens += int(ctx.get("metrics", {}).get("tokens_used", 0))
+    typer.echo(json.dumps({"session_id": session_id, "tokens_used": tokens}))
